♻️ Move identical pages to own component
D.R.Y.
