+++ src/pages/event/[slug].tsx
@@ -1,13 +1,8 @@
 import { ParsedUrlQuery } from 'querystring';
-import { useTimeout } from '@chakra-ui/react';
-import { differenceInMilliseconds, formatISO, parseISO } from 'date-fns';
 import { GetServerSideProps } from 'next';
-import { useRouter } from 'next/router';
 import React from 'react';
-import ErrorBox from '../../components/error-box';
-import HappeningUI from '../../components/happening';
-import SEO from '../../components/seo';
 import { Happening, HappeningAPI, HappeningType, RegistrationAPI, SpotRangeCount } from '../../lib/api';
+import HappeningPage from '../../components/happening-page';
 
 interface Props {
     happening: Happening | null;
@@ -18,36 +13,14 @@ interface Props {
 }
 
 const EventPage = ({ happening, backendUrl, spotRangeCounts, date, error }: Props): JSX.Element => {
-    const router = useRouter();
-    const regDate = parseISO(happening?.registrationDate ?? formatISO(new Date()));
-    const time =
-        !happening ||
-        differenceInMilliseconds(regDate, date) < 0 ||
-        differenceInMilliseconds(regDate, date) > 172_800_000
-            ? null
-            : differenceInMilliseconds(regDate, date);
-
-    /* eslint-disable @typescript-eslint/no-floating-promises */
-    useTimeout(() => {
-        if (happening?.registrationDate) router.replace(router.asPath, undefined, { scroll: false });
-    }, time);
-    /* eslint-enable @typescript-eslint/no-floating-promises */
-
     return (
-        <>
-            {error && !happening && <ErrorBox error={error} />}
-            {happening && !error && (
-                <>
-                    <SEO title={happening.title} />
-                    <HappeningUI
-                        happening={happening}
-                        backendUrl={backendUrl}
-                        spotRangeCounts={spotRangeCounts}
-                        date={date}
-                    />
-                </>
-            )}
-        </>
+        <HappeningPage
+            happening={happening}
+            backendUrl={backendUrl}
+            spotRangeCounts={spotRangeCounts}
+            date={date}
+            error={error}
+        />
     );
 };
